Split up `InstantExecutionIntegrationTest` into some separate classes.
